+++ mayan/apps/tags/tests/test_views.py
@@ -1,5 +1,7 @@
 from __future__ import unicode_literals
 
+from django.utils.encoding import force_text
+
 from mayan.apps.common.tests import GenericViewTestCase
 from mayan.apps.documents.permissions import permission_document_view
 from mayan.apps.documents.tests import GenericDocumentViewTestCase
@@ -100,8 +102,27 @@ class TagViewTestCase(TagTestMixin, TagViewTestMixin, GenericViewTestCase):
         self.test_tag.refresh_from_db()
         self.assertNotEqual(self.test_tag.label, tag_label)
 
+    def test_tag_list_view_with_no_permission(self):
+        self._create_test_tag()
+
+        response = self._request_test_tag_list_view()
+        self.assertNotContains(
+            response=response, text=self.test_tag.label, status_code=200
+        )
+
+    def test_tag_list_view_with_access(self):
+        self._create_test_tag()
+
+        self.grant_access(obj=self.test_tag, permission=permission_tag_view)
+
+        response = self._request_test_tag_list_view()
+        self.assertContains(
+            response=response, text=self.test_tag.label, status_code=200
+        )
+
 
 class TagDocumentViewTestCase(TagTestMixin, TagViewTestMixin, GenericDocumentViewTestCase):
+    """
     def test_document_tags_widget_no_permissions(self):
         self._create_test_tag()
 
@@ -126,6 +147,58 @@ class TagDocumentViewTestCase(TagTestMixin, TagViewTestMixin, GenericDocumentVie
         self.assertContains(
             response=response, text=self.test_tag.label, status_code=200
         )
+    """
+    def test_document_tags_list_no_permissions(self):
+        self._create_test_tag()
+
+        self.test_tag.documents.add(self.test_document)
+
+        response = self._request_test_document_tag_list_view()
+        self.assertNotContains(
+            response=response, text=force_text(self.test_tag), status_code=404
+        )
+
+    def test_document_tags_list_with_document_access(self):
+        self._create_test_tag()
+
+        self.test_tag.documents.add(self.test_document)
+
+        self.grant_access(
+            obj=self.test_document, permission=permission_tag_view
+        )
+
+        response = self._request_test_document_tag_list_view()
+        self.assertNotContains(
+            response=response, text=force_text(self.test_tag), status_code=200
+        )
+
+    def test_document_tags_list_with_tag_access(self):
+        self._create_test_tag()
+
+        self.test_tag.documents.add(self.test_document)
+
+        self.grant_access(obj=self.test_tag, permission=permission_tag_view)
+
+        response = self._request_test_document_tag_list_view()
+        self.assertNotContains(
+            response=response, text=force_text(self.test_tag), status_code=404
+        )
+
+    def test_document_tags_list_with_full_access(self):
+        self._create_test_tag()
+
+        self.test_tag.documents.add(self.test_document)
+
+        self.grant_access(obj=self.test_tag, permission=permission_tag_view)
+        self.grant_access(
+            obj=self.test_document, permission=permission_tag_view
+        )
+
+        response = self._request_test_document_tag_list_view()
+        self.assertContains(
+            response=response, text=force_text(self.test_tag), status_code=200
+        )
+
 
     def test_document_attach_tag_view_no_permission(self):
         self._create_test_tag()